% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{classification_summary_workflow}
\alias{classification_summary_workflow}
\title{Workflow to get all classification performance}
\usage{
classification_summary_workflow(cv_trained, show_feature_map = FALSE)
}
\arguments{
\item{cv_trained}{An output of \code{\link{cv_loop_train}} or
\code{\link{cv_loop_train_parallel}}}
}
\value{
A list of classification performance summary
}
\description{
Workflow to get all classification performance
}
