% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{combine.predicted.list}
\alias{combine.predicted.list}
\title{combine outcomes of multiple predictors}
\usage{
combine.predicted.list(predicted.list, min.prob = 0.25)
}
\arguments{
\item{predicted.list}{a list of predicted outcomes (tables).
Each table should have two columns: prob, predicted}

\item{min.prob}{a minimum probability/likelihood to be called a class.
(default = 0.25)}
}
\value{
combined table
        (prob, predicted) of each table will be prefixed with class name,
           and combined.  The last two columns will be the final outcomes.
}
\description{
combine outcomes of multiple predictors
}
