% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_batches}
\alias{combine_batches}
\title{Combine SummarizedExperiment objects and correct batch effect}
\usage{
combine_batches(
  se.list,
  ID = "ID",
  by.assay = "data",
  correct.batch = TRUE,
  ref.batch = NULL,
  visual.aid = FALSE,
  ...
)
}
\arguments{
\item{se.list}{a list of SummarizedExperiment objects}

\item{ID}{a column name or index for gene ID}

\item{by.assay}{assay}

\item{ref.batch}{passed to \code{\link{ComBat}}}

\item{visual.aid}{if TRUE, add UMAPs before and after batch correction}
}
\value{
combined SummarizedExperiment object
}
\description{
Combine SummarizedExperiment objects and correct batch effect
}
