% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_batches_dfs}
\alias{combine_batches_dfs}
\title{Combine SummarizedExperiment objects and correct batch effect}
\usage{
combine_batches_dfs(df.list, ID = "ID", ref.batch = NULL, correct.batch = TRUE)
}
\arguments{
\item{df.list}{a list of data frames}

\item{ID}{a column name or index for gene ID}

\item{ref.batch}{passed to \code{\link{ComBat}}}
}
\value{
combined data frame
}
\description{
Combine SummarizedExperiment objects and correct batch effect
}
