% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_train.R
\name{compute_difference}
\alias{compute_difference}
\title{A function to compute difference between two classes}
\usage{
compute_difference(dd, cls)
}
\arguments{
\item{dd}{A \code{matrix} or \code{data.frame}
where each row is an observation}

\item{cls}{A set of classes, either in \code{factor}}
}
\value{
A vector of difference
           \code{dd[cls == factor.1] - dd[cls==factor.2]}
}
\description{
A function to compute difference between two classes
}
