% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{compute_pca}
\alias{compute_pca}
\title{A function to perform *PCA* projection of data}
\usage{
compute_pca(se, assay = "data")
}
\arguments{
\item{se}{A \code{\link{SummarizedExperiment}} object}

\item{assay}{A name of the slot data is stored}
}
\value{
A \code{\link{SummarizedExperiment}} object with *PCA* projection
  stored in \code{metadata(se[[pca]]} slot
}
\description{
A function to perform *PCA* projection of data
}
