% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{compute_tsne}
\alias{compute_tsne}
\title{A function to perform *t-SNE* projection of data}
\usage{
compute_tsne(se, assay = "data")
}
\arguments{
\item{se}{A \code{\link{SummarizedExperiment}} object}

\item{assay}{A name of the slot data is stored}
}
\value{
A \code{\link{SummarizedExperiment}} object with *t-SNE* projection stored in \code{metadata(se[[tsne]]} slot
}
\description{
A function to perform *t-SNE* projection of data
}
