% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_visualization.R
\name{construct_feature_heatmaps}
\alias{construct_feature_heatmaps}
\title{An internal function to construct heatmaps from feature maps}
\usage{
construct_feature_heatmaps(cv_trained_summary)
}
\arguments{
\item{cv_trained_summary}{An intermediate struct that contains \code{feature_maps}
and \code{accuracy} slots}
}
\description{
An internal function to construct heatmaps from feature maps
}
