% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_TPM.R
\name{counts_to_tpm}
\alias{counts_to_tpm}
\title{Convert counts to transcripts per million (TPM).}
\usage{
counts_to_tpm(counts, featureLength, meanFragmentLength = rep(1, ncol(counts)))
}
\arguments{
\item{counts}{A numeric matrix of raw feature counts i.e.
fragments assigned to each gene.}

\item{featureLength}{A numeric vector with feature lengths.}

\item{meanFragmentLength}{A numeric vector with mean fragment lengths.}
}
\value{
tpm A numeric matrix normalized by library size and feature length.
}
\description{
Convert a numeric matrix of features (rows) and conditions (columns) with
raw feature counts to transcripts per million.
}
\details{
Lior Pachter. Models for transcript quantification from RNA-Seq.
   arXiv:1104.3889v2

   Wagner, et al. Measurement of mRNA abundance using RNA-seq data:
   RPKM measure is inconsistent among samples. Theory Biosci. 24 July 2012.
   doi:10.1007/s12064-012-0162-3

   Source for the code: https://gist.github.com/slowkow/c6ab0348747f86e2748b
}
