% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_train.R
\name{cv_loop_train_iter}
\alias{cv_loop_train_iter}
\title{A function train a model with CV}
\usage{
cv_loop_train_iter(
  ii,
  data,
  cls,
  fitControl,
  resampling_rate,
  n_features,
  filter_method,
  filter_threshold_score,
  filter_threshold_diff,
  filter_direction,
  feature_uniform,
  feature_weighted
)
}
\arguments{
\item{ii}{integer (probably not needed)}

\item{data}{input matrix, of dimension nobs x nvars; each row is an observation vector.
Since this is an input to \code{glmnet}, it should be the format that can be used
with \code{glmnet}}

\item{cls}{class labels}

\item{fitControl}{A list of training parameters.  See \code{\link{caret::trainControl}} for detail}

\item{resampling_rate}{...}

\item{n_features}{...}

\item{filter_method}{...}

\item{filter_threshold_score}{...}

\item{filter_threshold_diff}{...}

\item{filter_direction}{...}

\item{feature_uniform}{...}

\item{feature_weighted}{...}
}
\description{
A function train a model with CV
}
