% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_train.R
\name{cv_loop_train_parallel}
\alias{cv_loop_train_parallel}
\title{iterate training a model with CV, using \code{\link{future.apply}}}
\usage{
cv_loop_train_parallel(
  data,
  cls,
  fitControl,
  K = 25,
  resampling_rate = 0.8,
  n_features = NA,
  filter_method = c("two.sided", "greater", "less"),
  filter_direction = "less",
  filter_threshold_diff = 1,
  filter_threshold_score = 0.8,
  feature_weights = c("uniform", "weighted", "both")
)
}
\arguments{
\item{data}{input matrix, of dimension \code{nobs x nvars}; each row is an observation vector.
Since this is an input to \code{\link{glmnet}}, it should be the format that can be used
with \code{\link{glmnet}}}

\item{cls}{class labels}

\item{K}{...}

\item{resampling_rate}{...}

\item{n_features}{...}

\item{filter_method}{...}

\item{filter_direction}{...}

\item{filter_threshold_diff}{...}

\item{filter_threshold_score}{...}

\item{feature_weights}{...}
}
\description{
iterate training a model with CV, using \code{\link{future.apply}}
}
