% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_train.R
\name{cv_train_final}
\alias{cv_train_final}
\title{A function to train a *final* model with CV}
\usage{
cv_train_final(
  data,
  cls,
  fitControl,
  filter_method = c("ROC", "WILCOX"),
  filter_direction = c("two.sided", "greater", "less"),
  observation_weights = NULL,
  feature_weights = c("uniform", "weighted")
)
}
\arguments{
\item{data}{input matrix, of dimension nobs x nvars; each row is an observation vector.
Since this is an input to \code{glmnet}, it should be the format that can be used
with \code{glmnet}}

\item{cls}{class labels}

\item{fitControl}{A list of training parameters.  See \code{\link{caret::trainControl}} for detail}

\item{filter_method}{... c("ROC", "WILCOX")}

\item{filter_direction}{... c("two.sided", "greater", "less")}

\item{observation_weights}{...}

\item{feature_weights}{... c("uniform", "weighted")}
}
\value{
a trained model (final) - a list of training parameters (see above), fitted model, and selected features
}
\description{
A function to train a *final* model with CV
}
