% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract performance metrics from \code{performance_summary} slot}
\usage{
extract_metrics(p.summary)
}
\arguments{
\item{p.summary}{Either "train" or "test" slot of an output of \code{\link{models_performance_summary}}}

\item{slot}{An index}
}
\description{
Extract performance metrics from \code{performance_summary} slot
}
