% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{extract_metrics_all}
\alias{extract_metrics_all}
\title{Extract all performance metrics}
\usage{
extract_metrics_all(p.summary)
}
\arguments{
\item{p.summary}{Either "train" or "test" slot of an output of \code{\link{models_performance_summary}}}
}
\description{
Extract all performance metrics
}
