% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_visualization.R
\name{feature_boxplots}
\alias{feature_boxplots}
\title{A function to construct boxplots for selected features}
\usage{
feature_boxplots(dd, cls, df = data.frame(), order.by = "")
}
\arguments{
\item{dd}{A \code{matrix} or \code{data.frame}
where each row is an observation}

\item{cls}{A set of classes, either in \code{factor}}

\item{df}{fold difference, computed by \code{compute_difference}}

\item{order.by}{which feature to order boxplots, typically `mean` or `diff`}
}
\description{
A function to construct boxplots for selected features
}
