% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_train.R
\name{filter_features}
\alias{filter_features}
\title{A function for filtering-based feature selection, using Wilcox test as a metric}
\usage{
filter_features(
  dd,
  cls,
  threshold = 0.8,
  direction = c("two.sided", "greater", "less")
)
}
\arguments{
\item{dd}{A \code{matrix} or \code{data.frame}
where each row is an observation}

\item{cls}{A set of classes, either in \code{factor}}

\item{threshold}{A minimum value of (1 - p.value of Wilcox test) for filtering}

\item{direction}{Direction of AUC statistical test: "two.sided", "greater" or "less"}
}
\description{
A function for filtering-based feature selection, using Wilcox test as a metric
}
