% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{flatten.predict.multiclass}
\alias{flatten.predict.multiclass}
\title{flatten the outcomes of \code{\link{multilayer.predict.type}}}
\usage{
flatten.predict.multiclass(predicted, target_classes, final.class = "final")
}
\arguments{
\item{predicted}{an outcome of a \code{.predict....} function}

\item{target_classes}{a set of classes to be flattened}

\item{final.class}{the name (column heading) of final class outcomes, after flattening classes}
}
\value{
the last column is added with the multi-class labels (flatted)
}
\description{
flatten the outcomes of \code{\link{multilayer.predict.type}}
}
