% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{flatten.predict.multilayer}
\alias{flatten.predict.multilayer}
\title{flatten the outcomes of \code{\link{multilayer.predict.type}}}
\usage{
flatten.predict.multilayer(predicted)
}
\arguments{
\item{predicted}{an outcome of \code{\link{multilayer.predict.type}}}
}
\value{
the last column is added with the multi-class labels (flatted)
}
\description{
flatten the outcomes of \code{\link{multilayer.predict.type}}
}
