% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{get.subsamples}
\alias{get.subsamples}
\title{get subset of samples with 'predicted' = 'cls'}
\usage{
get.subsamples(pred, cls)
}
\arguments{
\item{pred}{a table with 'Sample_ID' and 'predicted' columns}

\item{cls}{target class label}
}
\value{
a list of 'Sample_ID's
}
\description{
get subset of samples with 'predicted' = 'cls'
}
