% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_train.R
\name{get_selected_features}
\alias{get_selected_features}
\title{A function to get selected features from a model, \code{fit}}
\usage{
get_selected_features(fit, features)
}
\arguments{
\item{fit}{...}

\item{features}{...}
}
\description{
A function to get selected features from a model, \code{fit}
}
