% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_visualization.R
\name{ggplot_train_trajectory}
\alias{ggplot_train_trajectory}
\title{Create line plots to show train trajectories}
\usage{
ggplot_train_trajectory(cv_model)
}
\arguments{
\item{cv_model}{an output of caret::train or glmnet::cv_glmnet}
}
\value{
\code{\link{ggplot2}} object
}
\description{
Create line plots to show train trajectories
}
