% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_TPM.R
\name{gtf_to_gene_length}
\alias{gtf_to_gene_length}
\title{A function to extract 'gene_length' from GTF data frame}
\usage{
gtf_to_gene_length(gtf_df)
}
\arguments{
\item{gtf_df}{data.frame formatted GTF file.  Can be obtained by:

    gtf <- rtracklayer::import('Homo_sapiens.GRCh38.94.gtf')
    gtf_df = as.data.frame(gtf)}
}
\value{
feature_lengths in data frame
}
\description{
A function to extract 'gene_length' from GTF data frame
}
