% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_visualization.R
\name{heatmap_features_cor}
\alias{heatmap_features_cor}
\title{Create a heatmap of correlation between features (rows)}
\usage{
heatmap_features_cor(dd, importance, ...)
}
\arguments{
\item{dd}{data (row: feature, column: observation)}

\item{importance}{a vector of `importance` values of features}
}
\value{
\code{\link{ComplexHeatmap}} object
}
\description{
Create a heatmap of correlation between features (rows)
}
