% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_visualization.R
\name{heatmap_glmnet}
\alias{heatmap_glmnet}
\title{Create a heatmap of projected data (coef x data) with predicted outcomes at the top and consensus}
\usage{
heatmap_glmnet(se, assay = "data", cls_results, cv_model, scale = FALSE, ...)
}
\arguments{
\item{se}{\code{\link{SummarizedExperiment}} container}

\item{assay}{a name of assay slot in \code{se}}

\item{cls_results}{An output of \code{\link{classification_summary_workflow}}}

\item{cv_model}{output of caret::train or glmnet::cv_glmnet}

\item{scale}{If TRUE, the data (=assays(se)[[assay]]) will be scaled}

\item{...}{Additional parameters to be passed to \code{\link{Heatmap}}}
}
\value{
Heatmap object
}
\description{
Create a heatmap of projected data (coef x data) with predicted outcomes at the top and consensus
}
