% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_visualization.R
\name{heatmap_prediction}
\alias{heatmap_prediction}
\title{Create a heatmap of data with predicted outcomes at the top and consensus}
\usage{
heatmap_prediction(se, assay = "data", cls_results, scale = FALSE, ...)
}
\arguments{
\item{se}{\code{\link{SummarizedExperiment}} container}

\item{assay}{a name of assay slot in \code{se}}

\item{cls_results}{An output of \code{\link{classification_summary_workflow}}}

\item{scale}{If TRUE, the data (=assays(se)[[assay]]) will be scaled}

\item{...}{Additional parameters to be passed to \code{\link{Heatmap}}}
}
\value{
Heatmap object
}
\description{
Create a heatmap of data with predicted outcomes at the top and consensus
}
