% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{models_performance_summary}
\alias{models_performance_summary}
\title{Generate a summary of performance of CV trained models}
\usage{
models_performance_summary(cv_trained)
}
\arguments{
\item{cv_trained}{An output of \code{\link{cv_loop_train}} or \code{\link{cv_loop_train_parallel}}}
}
\description{
Generate a summary of performance of CV trained models
}
