% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_train.R
\name{mult_wilcox_test}
\alias{mult_wilcox_test}
\title{Internal function for multiple Wilcox test}
\usage{
mult_wilcox_test(dd, cls, ...)
}
\arguments{
\item{dd}{A \code{matrix} or \code{data.frame}
where each row is an observation}

\item{cls}{A set of classes, either in \code{factor}}
}
\description{
only works for binary
  cls should be 0 or 1
}
