% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{predict.consensus.type}
\alias{predict.consensus.type}
\title{convert the outcomes of \code{\link{predict.consensus}}
  to the one equivalent of \code{\link{predict.type}}}
\usage{
\method{predict}{consensus.type}(cpred)
}
\arguments{
\item{cpred}{an outcome of \code{\link{predict.consensus}}}
}
\value{
the last column is added with the multi-class labels (flatted)
}
\description{
convert the outcomes of \code{\link{predict.consensus}}
  to the one equivalent of \code{\link{predict.type}}
}
