% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{predict.multiClass}
\alias{predict.multiClass}
\title{apply a list of models to a set of samples}
\usage{
\method{predict}{multiClass}(models, dd, type = "raw")
}
\arguments{
\item{models}{a list of models which is an outcome of
\code{\link{cv_loop_train_iter}} or \code{\link{cv_train_final}}, a model
trained for a class}

\item{dd}{input matrix, of dimension \code{nobs x nvars}; each row is an
observation vector. Since this is an input to \code{\link{glmnet}}, it
should be the format that can be used with \code{\link{glmnet}}}

\item{type}{either "raw" or "prob", for the number/class predictions or class
probabilities, respectively. Class probabilities are not available for all
classification models}
}
\value{
a table where each rows is predicted outcomes of applying predictive
  models to a sample
}
\description{
returns a table where each rows is predicted outcomes of a sample, by
applying predictive models
}
