% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{predict.multilayer}
\alias{predict.multilayer}
\title{predict outcomes of observations in data using *selected features*}
\usage{
\method{predict}{multilayer}(models, dd)
}
\arguments{
\item{dd}{input matrix, of dimension \code{nobs x nvars}; each row is an observation vector.
Since this is an input to \code{\link{glmnet}}, it should be the format that can be used
with \code{\link{glmnet}}}

\item{model}{a list of models which is an outcome of \code{\link{cv_loop_train_iter}} or
 \code{\link{cv_train_final}}, a model trained for a class
see \code{example} below.}

\item{type}{'raw' (default) or 'prob'}
}
\value{
a named vector, either outputs class labels ('raw) or
        probability tables, a column for each class label ('prob')

                 mm = list(
                   main = a_main_model,
                   c1 = a_c1_model,
                   c2 = a_c2_model
                 )

                 main is applied first which should classify samples into either c1 or c2
                 all samples classified to c1 will be then fed into a_c1_model, and
                 the samples classified to c2 will be fed into a_c2_model.

                 a model can be NULL, meaning no more action.

                 The list can be nested.  For example,

                 mm = list(
                   main = a_main_model,
                   c1 = a_c1_model,
                   c2 = list (
                     main = a_c2_model,
                     c2a = a_c2a_model,
                     c2b = a_c2b_model
                   )
                 )
}
\description{
predict outcomes of observations in data using *selected features*
}
