% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{predict.type.subclass}
\alias{predict.type.subclass}
\title{sub-classify a subset of samples belonging to a class}
\usage{
\method{predict}{type.subclass}(predicted, target_column, subclassify, submodel, data)
}
\arguments{
\item{predicted}{a table where each rows is predicted outcomes of a sample,
i.e. outcome of a \code{predict...} function}

\item{target_column}{which column to look for class label}

\item{subclassify}{a class to subclassify}

\item{submodel}{a model which is an outcome of \code{\link{cv_loop_train_iter}}
or \code{\link{cv_final_train}}}

\item{data}{input matrix, of dimension \code{nobs x nvars}; each row is an observation vector.
Since this is an input to \code{\link{glmnet}}, it should be the format that can be used
with \code{\link{glmnet}}}
}
\value{
amended predicted table with subclass labels
}
\description{
returns a table where each rows is predicted outcomes of a sample,
     by applying a predictive model
}
