% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{prediction_consensus_summary}
\alias{prediction_consensus_summary}
\title{Generate a consensus summary of outcomes}
\usage{
prediction_consensus_summary(cv_trained)
}
\arguments{
\item{cv_trained}{An output of \code{\link{cv_loop_train}} or
\code{\link{cv_loop_train_parallel}}}
}
\value{
A table of consensus summary of outcomes
}
\description{
Generate a consensus summary of outcomes
}
