% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{prediction_summary}
\alias{prediction_summary}
\title{Generate a list of prediction summary}
\usage{
prediction_summary(pred_tbls)
}
\arguments{
\item{pred_tbls}{An output from \code{\link{prediction_tables}}}
}
\value{
A list of prediction summary
}
\description{
Generate a list of prediction summary
}
