% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_model_summary.R
\name{prediction_tables}
\alias{prediction_tables}
\title{predict outcomes of observations in data using *selected features*}
\usage{
prediction_tables(data, cls, models)
}
\arguments{
\item{data}{input matrix, of dimension \code{nobs x nvars}; each row is an observation vector.
Since this is an input to \code{\link{glmnet}}, it should be the format that can be used
with \code{\link{glmnet}}}

\item{cls}{class labels}

\item{model}{a CV trained model, an output of cv_loop_train_iter}
}
\value{
A list of
    class (class labels),
    trained (names of trained samples),
    row.names (selected features),
    pred (predicted outcomes), and
    prob (probabilities of predicted outcomes)
}
\description{
predict outcomes of observations in data using *selected features*
}
