% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantize.R
\name{quantize_robust_zscore}
\alias{quantize_robust_zscore}
\title{Quantize values in a matrix}
\usage{
quantize_robust_zscore(rz)
}
\arguments{
\item{rz}{robust z-score matrix}
}
\value{
quantized robust z-score matrix
}
\description{
Quantize values in a matrix
}
