% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{read_gmt}
\alias{read_gmt}
\title{A function to read GMT file}
\usage{
read_gmt(
  gmt.file,
  min.size = 1,
  max.size = -1,
  simplify = TRUE,
  sep1 = "\\t",
  sep2 = sep1
)
}
\arguments{
\item{gmt.file}{a text file in 'gmt' format.}

\item{min.size}{The minimum gene set size to read.  Defaults to 0.}

\item{max.size}{The maximum gene set size to read.  Defaults to no limit.}

\item{simplify}{Defaults to TRUE.  See 'output' for detail.}

\item{sep1}{a character string separating 'gene set name', 'url' and 'gene set'}

\item{sep2}{a character string separating each gene in 'gene set'}
}
\value{
a list of gene set.
       If TRUE, the return structure is simplified to a list of genesets
       each of which is a list of genes and the name of each geneset is
       the name of gene set from MSigDB file, the first element of each line.
       If FALSE, it returns a list of three element list each element of which
       is \code{name}, \code{url}, and \code{geneset}.
}
\description{
A function to read GMT file
}
