% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexHeatmap_utils.R
\name{save.PDF.Heatmap}
\alias{save.PDF.Heatmap}
\title{A function to save a Heatmap object to a PDF file}
\usage{
save.PDF.Heatmap(
  ht,
  filename,
  width = width.Heatmap(ht, "inch"),
  height = height.Heatmap(ht, "inch")
)
}
\description{
A function to save a Heatmap object to a PDF file
}
