% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexHeatmap_utils.R
\name{size.Heatmap}
\alias{size.Heatmap}
\title{A function to estimate the drawing size of Heatmap object}
\usage{
size.Heatmap(ht, unitTo = "inch")
}
\arguments{
\item{ht}{Heatmap object, an output of draw(a_Heatmap_object)
and \code{accuracy} slots}

\item{unitTo}{The coordinate system to convert the unit to. See the unit function for valid coordinate systems,
to be passed to grid::convertHeight() and grid::convertWidth() functions.}
}
\value{
a list of width and height in "unitTo"
}
\description{
A function to estimate the drawing size of Heatmap object
}
