% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_se}
\alias{split_se}
\title{Split SummarizedExperiment object}
\usage{
split_se(se, col)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{col}{integer (index) or string (column heading) of colData(se)}
}
\value{
a list of split objects
}
\description{
Split SummarizedExperiment object
}
