% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_visualization.R
\name{construct_feature_maps}
\alias{construct_feature_maps}
\title{Construct feature maps}
\usage{
construct_feature_maps(models)
}
\arguments{
\item{models}{a list of models each of which is an output of \code{\link{cv_loop_train_iter}}}
}
\value{
a list ("uniform" and/or "weighted") of feature map which is made of
        * columns of filtered features, amended by \code{overall_score}
        * predictors (selected by model), amended by \code{overall_score, overall_feature_score}
        * model coefficients, amended by \code{overall_score, overall_score_mad, overall_score2, overall_feature_score}
}
\description{
Construct feature maps
}
